Add files via upload
Create a Web Service for translating between two languages of your choice. The client requires a specific word, checking the service more accurately in the XML database. If the word exists, the service broadcasts the translation - if it does not broadcast the message. The client inputs the parameters: the word for translation, the source language, the target language.